requirements.txt
flask
flask-cors
# transformers
# torch
tensorflow-gpu==1.14
gpt-2-simple
